@@ -1,15 +1,19 @@
+import json
 import time
-from typing import Dict, List
 
 import pytest
 
 from ae5_tools.api import AEUserSession
-from tests.utils import _cmd, _get_vars
+from tests.adsp.common.utils import _cmd, _get_vars
+from tests.system.state import load_account
 
 
 @pytest.fixture(scope="session")
 def user_session():
-    hostname, username, password = _get_vars("AE5_HOSTNAME", "AE5_USERNAME", "AE5_PASSWORD")
+    hostname: str = _get_vars("AE5_HOSTNAME")
+    local_account: dict = load_account(id="1")
+    username: str = local_account["username"]
+    password: str = local_account["password"]
     s = AEUserSession(hostname, username, password)
     yield s
     s.disconnect()
@@ -30,12 +34,10 @@ def test_job_run(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
-        "job", "create", prec["id"], "--name", "testjob1", "--command", "run", "--run", "--wait"
-    )
+    create_job_result: dict = _cmd("job", "create", prec["id"], "--name", "testjob1", "--command", "run", "--run", "--wait")
 
     # Execute the test (Run a previously created job)
-    run_job_result: Dict = _cmd("job", "run", "testjob1")
+    run_job_result: dict = _cmd("job", "run", "testjob1")
 
     # Review Test Results
     assert run_job_result["name"] == "testjob1"
@@ -47,7 +49,7 @@ def test_job_run(cli_project):
     max_loop: int = 100
     wait: bool = True
     while wait:
-        run_once_status: Dict = _cmd("run", "info", run_job_result["id"])
+        run_once_status: dict = _cmd("run", "info", run_job_result["id"])
         if run_once_status["state"] == "completed":
             wait = False
         else:
@@ -60,7 +62,7 @@ def test_job_run(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
@@ -78,9 +80,7 @@ def test_job_run_implicit_revision_latest(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
-        "job", "create", prec["id"], "--name", "testjob1", "--command", "run", "--run", "--wait"
-    )
+    create_job_result: dict = _cmd("job", "create", prec["id"], "--name", "testjob1", "--command", "run", "--run", "--wait")
 
     # Review Test Results
     assert create_job_result["name"] == "testjob1"
@@ -90,7 +90,7 @@ def test_job_run_implicit_revision_latest(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
@@ -103,9 +103,7 @@ def test_job_run_explicit_revision_latest(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
-        "job", "create", f"{prec['id']}:latest", "--name", "testjob1", "--command", "run", "--run", "--wait"
-    )
+    create_job_result: dict = _cmd("job", "create", f"{prec['id']}:latest", "--name", "testjob1", "--command", "run", "--run", "--wait")
 
     # Review Test Results
     assert create_job_result["name"] == "testjob1"
@@ -115,7 +113,7 @@ def test_job_run_explicit_revision_latest(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
@@ -128,9 +126,7 @@ def test_job_run_explicit_revision_first(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
-        "job", "create", f"{prec['id']}:0.1.0", "--name", "testjob1", "--command", "run", "--run", "--wait"
-    )
+    create_job_result: dict = _cmd("job", "create", f"{prec['id']}:0.1.0", "--name", "testjob1", "--command", "run", "--run", "--wait")
 
     # Review Test Results
     assert create_job_result["name"] == "testjob1"
@@ -140,7 +136,7 @@ def test_job_run_explicit_revision_first(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
@@ -158,9 +154,7 @@ def test_job_run_by_owner_and_name_implicit_revision_latest(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
-        "job", "create", f"{prec['owner']}/{prec['name']}", "--name", "testjob1", "--command", "run", "--run", "--wait"
-    )
+    create_job_result: dict = _cmd("job", "create", f"{prec['owner']}/{prec['name']}", "--name", "testjob1", "--command", "run", "--run", "--wait")
 
     # Review Test Results
     assert create_job_result["name"] == "testjob1"
@@ -170,7 +164,7 @@ def test_job_run_by_owner_and_name_implicit_revision_latest(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
@@ -183,7 +177,7 @@ def test_job_run_by_owner_and_name_explicit_revision_latest(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
+    create_job_result: dict = _cmd(
         "job",
         "create",
         f"{prec['owner']}/{prec['name']}:latest",
@@ -203,7 +197,7 @@ def test_job_run_by_owner_and_name_explicit_revision_latest(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
@@ -216,7 +210,7 @@ def test_job_run_by_owner_and_name_explicit_revision_first(cli_project):
 
     # Create a pre-existing job, (run it and wait for completion)
     prec = cli_project
-    create_job_result: Dict = _cmd(
+    create_job_result: dict = _cmd(
         "job",
         "create",
         f"{prec['owner']}/{prec['name']}:0.1.0",
@@ -236,7 +230,7 @@ def test_job_run_by_owner_and_name_explicit_revision_first(cli_project):
     # Cleanup after the test
 
     # Remove runs
-    job_runs: List[Dict] = _cmd("job", "runs", create_job_result["id"])
+    job_runs: list[dict] = _cmd("job", "runs", create_job_result["id"])
     for run in job_runs:
         _cmd("run", "delete", run["id"])
 
