@@ -1,15 +1,19 @@
+import json
 import uuid
-from typing import List
 
 import pytest
 
 from ae5_tools.api import AEUserSession
-from tests.utils import CMDException, _cmd, _get_vars
+from tests.adsp.common.utils import CMDException, _cmd, _get_vars
+from tests.system.state import load_account
 
 
 @pytest.fixture(scope="session")
 def user_session():
-    hostname, username, password = _get_vars("AE5_HOSTNAME", "AE5_USERNAME", "AE5_PASSWORD")
+    hostname: str = _get_vars("AE5_HOSTNAME")
+    local_account: dict = load_account(id="1")
+    username: str = local_account["username"]
+    password: str = local_account["password"]
     s = AEUserSession(hostname, username, password)
     yield s
     s.disconnect()
@@ -30,7 +34,7 @@ def test_secret_create_and_list_and_delete(user_session, secret_name):
     secret_add_result: str = _cmd("secret", "add", secret_name, secret_value)
     assert secret_add_result == ""
 
-    user_secrets: List[str] = user_session.secret_list()
+    user_secrets: list[dict] = user_session.secret_list()
     assert secret_name in [secret["secret_name"] for secret in user_secrets]
 
     secret_delete_result: str = _cmd("secret", "delete", secret_name)
