@@ -1,15 +1,18 @@
 import uuid
-from typing import List
 
 import pytest
 
 from ae5_tools.api import AEException, AEUserSession
-from tests.utils import _get_vars
+from tests.adsp.common.utils import _get_vars
+from tests.system.state import load_account
 
 
 @pytest.fixture(scope="session")
 def user_session():
-    hostname, username, password = _get_vars("AE5_HOSTNAME", "AE5_USERNAME", "AE5_PASSWORD")
+    hostname: str = _get_vars("AE5_HOSTNAME")
+    local_account: dict = load_account(id="1")
+    username: str = local_account["username"]
+    password: str = local_account["password"]
     s = AEUserSession(hostname, username, password)
     yield s
     s.disconnect()
@@ -27,7 +30,7 @@ def secret_name():
 
 def test_secret_add_and_list(user_session, secret_name):
     user_session.secret_add(key=secret_name, value=str(uuid.uuid4()))
-    results: List[str] = user_session.secret_list()
+    results: list[dict] = user_session.secret_list()
     assert secret_name in [secret["secret_name"] for secret in results]
 
 
