@@ -115,16 +115,12 @@ def _start(**kwargs):
     required=False,
     help="Deployment name. If not supplied, it is autogenerated from the project name.",
 )
-@click.option(
-    "--endpoint", type=str, required=False, help="Endpoint name. If not supplied, a generated subdomain will be used."
-)
+@click.option("--endpoint", type=str, required=False, help="Endpoint name. If not supplied, a generated subdomain will be used.")
 @click.option("--command", help="The command to use for this deployment.")
 @click.option("--resource-profile", help="The resource profile to use for this deployment.")
 @click.option("--public/--private", is_flag=True, help="Make the deployment public/private (default).")
 @click.option("--wait", is_flag=True, help="Wait for the deployment to complete initialization before exiting.")
-@click.option(
-    "--stop-on-error", is_flag=True, help="Stop the deployment if it fails on the first attempt. Implies --wait."
-)
+@click.option("--stop-on-error", is_flag=True, help="Stop the deployment if it fails on the first attempt. Implies --wait.")
 @click.option("--open", is_flag=True, help="Open a browser upon initialization. Implies --wait.")
 @click.option("--frame", is_flag=True, help="Include the AE banner when opening.")
 @global_options
@@ -149,9 +145,7 @@ def start(**kwargs):
 @deployment.command()
 @ident_filter("deployment", required=True)
 @click.option("--wait", is_flag=True, help="Wait for the deployment to complete initialization before exiting.")
-@click.option(
-    "--stop-on-error", is_flag=True, help="Stop the deployment if it fails on the first attempt. Implies --wait."
-)
+@click.option("--stop-on-error", is_flag=True, help="Stop the deployment if it fails on the first attempt. Implies --wait.")
 @click.option("--open", is_flag=True, help="Open a browser upon initialization. Implies --wait.")
 @click.option("--frame", is_flag=True, help="Include the AE banner when opening.")
 @global_options
@@ -161,9 +155,7 @@ def restart(**kwargs):
     The DEPLOYMENT identifier need not be fully specified, and may even include
     wildcards. But it must match exactly one project.
     """
-    result = cluster_call(
-        "deployment_restart", **kwargs, prefix="Restarting deployment {ident}...", postfix="restarted."
-    )
+    result = cluster_call("deployment_restart", **kwargs, prefix="Restarting deployment {ident}...", postfix="restarted.")
 
 
 @deployment.command(short_help="Stop a deployment.")
