@@ -1,12 +1,18 @@
+import json
+
 import pytest
 
 from ae5_tools.api import AEUserSession
-from tests.utils import _cmd, _get_vars
+from tests.adsp.common.utils import _cmd, _get_vars
+from tests.system.state import load_account
 
 
 @pytest.fixture(scope="session")
 def user_session():
-    hostname, username, password = _get_vars("AE5_HOSTNAME", "AE5_USERNAME", "AE5_PASSWORD")
+    hostname: str = _get_vars("AE5_HOSTNAME")
+    local_account: dict = load_account(id="1")
+    username: str = local_account["username"]
+    password: str = local_account["password"]
     s = AEUserSession(hostname, username, password)
     yield s
     s.disconnect()
