@@ -1,6 +1,6 @@
 import pytest
 
-from .utils import _cmd
+from tests.adsp.common.utils import _cmd
 
 
 @pytest.fixture(scope="module")
